feat: attachments crud

+++ modules/Content/Http/Controllers/AttachmentController.php
@@ -2,7 +2,7 @@
 
 namespace Modules\Content\Http\Controllers;
 use App\Http\Controllers\Controller;
-
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
 use Modules\Content\Models\Attachment;
 use Illuminate\Support\Facades\Auth;
@@ -19,7 +19,7 @@ class AttachmentController extends Controller
   public function index(Request $request)
   {
     $attachments = Attachment::where('site_id', $request->site->id);
-    if(Auth::user()->enrol) {
+    if(Auth::user()->enrollment) {
       $attachments = $attachments->where('user_id', $request->user_id);
     }
     if($request->has('search')) {
@@ -36,17 +36,17 @@ class AttachmentController extends Controller
   /**
    * Store a newly created resource in storage.
    */
-  public function store($siteId, Request $request)
+  public function store(Request $request)
   {
     $request->validate([
-      'file' => 'required|file|max:1024|mimes:gif,png,jpg,jpeg,pdf,doc,docx,xls,xlsx,ppt,pptx',
+      'file' => 'required|file|max:5120|mimes:gif,png,jpg,jpeg,pdf,doc,docx,xls,xlsx,ppt,pptx',
       'width' => 'integer',
       'height' => 'integer',
     ]);
 
     $attachment = new Attachment();
     $attachment->user_id = Auth::user()->id;
-    $attachment->site_id = $siteId;
+    $attachment->site_id = $request->site->id;
     $attachment->name = $request->file('file')->getClientOriginalName();
     // transform image
     if($request->width > 0 && $request->height > 0) {
@@ -56,12 +56,13 @@ class AttachmentController extends Controller
       $randomStr = Str::random(40).".webp";
       Storage::disk('local')->put('tmp/'.$randomStr, $imgfinal);
       $imgfinal = new File(Storage::path('tmp/'.$randomStr));
-      $attachment->path = Storage::disk('digitalocean')->putFile($siteId.'/'.Auth::user()->id, $imgfinal, 'public');
+      $attachment->path = Storage::disk('digitalocean')->putFile($request->site->id.'/'.Auth::user()->id, $imgfinal, 'public');
       // Delete temporary file
       Storage::disk('local')->delete('tmp/'.$randomStr);
     } else {
-      $attachment->path = Storage::disk('digitalocean')->putFile($siteId.'/'.Auth::user()->id, $request->file('file'), 'public');
+      $attachment->path = Storage::disk('digitalocean')->putFile($request->site->id.'/'.Auth::user()->id, $request->file('file'), 'public');
     }
+    $attachment->provider = 'digitalocean';
     $attachment->size = $request->file('file')->getSize();
     $attachment->extension = $request->file('file')->extension();
     $attachment->mime_type = $request->file('file')->getMimeType();
@@ -73,9 +74,11 @@ class AttachmentController extends Controller
   /**
    * Remove the specified resource from storage.
    */
-  public function destroy($siteId, string $id)
+  public function destroy(Request $request, string $id)
   {
-    $attachment = Attachment::where('site_id', $siteId)->findOrFail($id);
+    $attachment = Attachment::where('site_id', $request->site->id)->findOrFail($id);
+    if(Gate::denies('manage-attachment', $attachment)) { return response()->json(['message' => 'Unauthorized'], 403); }
+
     Storage::disk('digitalocean')->delete($attachment->path);
     $attachment->delete();
